HTML code in a file named index.html and the CSS code in a file named styles.css.HTML file and the CSS file in the same directory.
Save the HTML code in a file named index.html and the CSS code in a file named styles.css. Make sure to place the HTML file and the CSS file in the same directory. Then, you can open the index.html file in a web browser to see the fruit menu with the applied styles.